+++ client/next.config.mjs
@@ -7,6 +7,7 @@ const nextConfig = {
       config.resolve.fallback = {
         ...config.resolve.fallback,
         ...webpackFallback,
+        fs: false,
       };
     }
