+++ client/src/modules/auth/lib/services/supabase/actions.ts
@@ -3,7 +3,7 @@
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
-import { createClient } from "./supabase/server";
+import { createClient } from "./server";
 
 type Credentials = {
   email: string;
